added cart functionality and refactoring

+++ src/main/java/kz/timshowtime/onlineShop/model/manyToMany/Embedded/CartItemId.java
@@ -7,5 +7,5 @@ import java.io.Serializable;
 @EqualsAndHashCode
 public class CartItemId implements Serializable {
     private int cart;
-    private int item;
+    private long item;
 }
